place counter into it's own component
